+++ opendc-compute/opendc-compute-service/src/main/kotlin/org/opendc/compute/service/driver/telemetry/GuestSystemStats.kt
@@ -35,5 +35,5 @@ import java.time.Instant
 public data class GuestSystemStats(
     val uptime: Duration,
     val downtime: Duration,
-    val bootTime: Instant
+    val bootTime: Instant?
 )
