+++ opendc-compute/opendc-compute-simulator/src/main/kotlin/org/opendc/compute/simulator/internal/Guest.kt
@@ -22,12 +22,6 @@
 
 package org.opendc.compute.simulator.internal
 
-import io.opentelemetry.api.common.AttributeKey
-import io.opentelemetry.api.common.Attributes
-import io.opentelemetry.api.common.AttributesBuilder
-import io.opentelemetry.api.metrics.ObservableDoubleMeasurement
-import io.opentelemetry.api.metrics.ObservableLongMeasurement
-import io.opentelemetry.semconv.resource.attributes.ResourceAttributes
 import kotlinx.coroutines.*
 import mu.KotlinLogging
 import org.opendc.compute.api.Server
@@ -75,11 +69,6 @@ internal class Guest(
      * a server.
      */
     var state: ServerState = ServerState.TERMINATED
-
-    /**
-     * The attributes of the guest.
-     */
-    val attributes: Attributes = GuestAttributes(this)
 
     /**
      * Start the guest.
@@ -158,7 +147,7 @@ internal class Guest(
         return GuestSystemStats(
             Duration.ofMillis(_uptime),
             Duration.ofMillis(_downtime),
-            Instant.ofEpochMilli(_bootTime)
+            _bootTime
         )
     }
 
@@ -235,7 +224,7 @@ internal class Guest(
      * This method is invoked when the guest was started on the host and has booted into a running state.
      */
     private fun onStart() {
-        _bootTime = clock.millis()
+        _bootTime = clock.instant()
         state = ServerState.RUNNING
         listener.onStart(this)
     }
@@ -250,18 +239,11 @@ internal class Guest(
         listener.onStop(this)
     }
 
-    private val STATE_KEY = AttributeKey.stringKey("state")
-
     private var _uptime = 0L
     private var _downtime = 0L
-    private val _upState = attributes.toBuilder()
-        .put(STATE_KEY, "up")
-        .build()
-    private val _downState = attributes.toBuilder()
-        .put(STATE_KEY, "down")
-        .build()
-
     private var _lastReport = clock.millis()
+    private var _bootTime: Instant? = null
+    private val _cpuLimit = machine.model.cpus.sumOf { it.frequency }
 
     /**
      * Helper function to track the uptime and downtime of the guest.
@@ -276,123 +258,5 @@ internal class Guest(
         } else if (state == ServerState.ERROR) {
             _downtime += duration
         }
-    }
-
-    /**
-     * Helper function to track the uptime of the guest.
-     */
-    fun collectUptime(result: ObservableLongMeasurement) {
-        updateUptime()
-
-        result.record(_uptime, _upState)
-        result.record(_downtime, _downState)
-    }
-
-    private var _bootTime = Long.MIN_VALUE
-
-    /**
-     * Helper function to track the boot time of the guest.
-     */
-    fun collectBootTime(result: ObservableLongMeasurement) {
-        if (_bootTime != Long.MIN_VALUE) {
-            result.record(_bootTime, attributes)
-        }
-    }
-
-    private val _activeState = attributes.toBuilder()
-        .put(STATE_KEY, "active")
-        .build()
-    private val _stealState = attributes.toBuilder()
-        .put(STATE_KEY, "steal")
-        .build()
-    private val _lostState = attributes.toBuilder()
-        .put(STATE_KEY, "lost")
-        .build()
-    private val _idleState = attributes.toBuilder()
-        .put(STATE_KEY, "idle")
-        .build()
-
-    /**
-     * Helper function to track the CPU time of a machine.
-     */
-    fun collectCpuTime(result: ObservableLongMeasurement) {
-        val counters = machine.counters
-        counters.flush()
-
-        result.record(counters.cpuActiveTime / 1000, _activeState)
-        result.record(counters.cpuIdleTime / 1000, _idleState)
-        result.record(counters.cpuStealTime / 1000, _stealState)
-        result.record(counters.cpuLostTime / 1000, _lostState)
-    }
-
-    private val _cpuLimit = machine.model.cpus.sumOf { it.frequency }
-
-    /**
-     * Helper function to collect the CPU limits of a machine.
-     */
-    fun collectCpuLimit(result: ObservableDoubleMeasurement) {
-        result.record(_cpuLimit, attributes)
-    }
-
-    /**
-     * An optimized [Attributes] implementation.
-     */
-    private class GuestAttributes(private val uid: String, private val attributes: Attributes) : Attributes by attributes {
-        /**
-         * Construct a [GuestAttributes] instance from a [Guest].
-         */
-        constructor(guest: Guest) : this(
-            guest.server.uid.toString(),
-            Attributes.builder()
-                .put(ResourceAttributes.HOST_NAME, guest.server.name)
-                .put(ResourceAttributes.HOST_ID, guest.server.uid.toString())
-                .put(ResourceAttributes.HOST_TYPE, guest.server.flavor.name)
-                .put(AttributeKey.longKey("host.num_cpus"), guest.server.flavor.cpuCount.toLong())
-                .put(AttributeKey.longKey("host.mem_capacity"), guest.server.flavor.memorySize)
-                .put(AttributeKey.stringArrayKey("host.labels"), guest.server.labels.map { (k, v) -> "$k:$v" })
-                .put(ResourceAttributes.HOST_ARCH, ResourceAttributes.HostArchValues.AMD64)
-                .put(ResourceAttributes.HOST_IMAGE_NAME, guest.server.image.name)
-                .put(ResourceAttributes.HOST_IMAGE_ID, guest.server.image.uid.toString())
-                .build()
-        )
-
-        override fun <T : Any?> get(key: AttributeKey<T>): T? {
-            // Optimize access to the HOST_ID key which is accessed quite often
-            if (key == ResourceAttributes.HOST_ID) {
-                @Suppress("UNCHECKED_CAST")
-                return uid as T?
-            }
-            return attributes.get(key)
-        }
-
-        override fun toBuilder(): AttributesBuilder {
-            val delegate = attributes.toBuilder()
-            return object : AttributesBuilder {
-
-                override fun putAll(attributes: Attributes): AttributesBuilder {
-                    delegate.putAll(attributes)
-                    return this
-                }
-
-                override fun <T : Any?> put(key: AttributeKey<Long>, value: Int): AttributesBuilder {
-                    delegate.put<T>(key, value)
-                    return this
-                }
-
-                override fun <T : Any?> put(key: AttributeKey<T>, value: T): AttributesBuilder {
-                    delegate.put(key, value)
-                    return this
-                }
-
-                override fun build(): Attributes = GuestAttributes(uid, delegate.build())
-            }
-        }
-
-        override fun equals(other: Any?): Boolean = attributes == other
-
-        // Cache hash code
-        private val _hash = attributes.hashCode()
-
-        override fun hashCode(): Int = _hash
     }
 }
