+++ settings.gradle.kts
@@ -49,7 +49,6 @@ include(":opendc-simulator:opendc-simulator-network")
 include(":opendc-simulator:opendc-simulator-compute")
 include(":opendc-telemetry:opendc-telemetry-api")
 include(":opendc-telemetry:opendc-telemetry-sdk")
-include(":opendc-telemetry:opendc-telemetry-compute")
 include(":opendc-trace:opendc-trace-api")
 include(":opendc-trace:opendc-trace-gwf")
 include(":opendc-trace:opendc-trace-swf")
